# lumina-api/swagger.yaml
openapi: 3.0.3
info:
  title: Lumina Bot API
  version: 3.0.0
  contact:
    name: Lumina Dev Team
    url: https://bot.luminasink.com
  description: |
    # Lumina Bot API

    REST API for the Lumina Bot dashboard, inventory/gacha system, Discord guild
    management, and the internal bot ↔ API bridge. Built on Node.js/Express,
    deployed to Vercel (production) and `localhost:3000` (dev).

    ---

    ## Authentication

    The API supports **two** authentication methods for user-facing routes:

    1. **Cookie `lumina_token` (preferred for browsers)**
       - `httpOnly`, `Secure` (prod), `SameSite=Lax`, max-age = 1 hour
       - Set by `POST /expapi/v1/login` and the OAuth2 callback
       - Sent automatically by the browser — **immune to XSS** (JavaScript cannot read it)

    2. **Header `Authorization: Bearer <jwt>` (legacy / non-browser)**
       - For CLI tools, scripts, integration tests, and any client that does not
         manage cookies
       - Same JWT, same 1h expiry

    JWT expiry: **1 hour**. There is no refresh-token rotation yet — clients must
    re-authenticate when the token expires.

    The session helper `verifyRequestAuth(req)` reads the cookie first, then
    falls back to the Bearer header. Either is accepted on every protected route.

    ---

    ## CSRF Protection

    State-changing routes (`POST`/`PUT`/`DELETE` with `csrfProtectionNeeded: true`)
    require the **double-submit cookie** pattern:

...

...
    rotated hourly.

...

    ---

...

...

...

    ---
...
servers:
...
components:
  securitySchemes:

    cookieAuth:
      type: apiKey
      in: cookie
      name: lumina_token
      description: |
        httpOnly JWT cookie set by `POST /expapi/v1/login` or the OAuth2 callback.
        Preferred auth method for browsers (immune to XSS). Sent automatically by
        the browser; non-browser clients should use `bearerAuth` instead.

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `Authorization: Bearer <jwt>` — legacy/alternative auth for non-browser
        clients (CLI, scripts, integration tests). Same JWT, same 1h expiry as
        the cookie.

    internalKey:
      type: apiKey
      in: header
      name: internal-key
      description: |
        Required on `/expapi/internal/*` routes. Constant-time compared against
        `INTERNAL_API_KEY`. Only the Discord bot should possess this key.

    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Required on state-changing routes (`POST`/`PUT`/`DELETE`) that have
        `csrfProtectionNeeded: true`. The value must match the `csrf_token`
        cookie obtained from `GET /expapi/v1/csrf-token`.

    apiKeyQuery:
      type: apiKey
      in: query
      name: apiKey
...
  schemas:
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable message (PT-BR).
        code:
          type: string
          description: Stable machine-readable error code.
          example: INVALID_CREDENTIALS
        details:
          type: object
          additionalProperties: true
          description: Optional structured details (e.g. attemptsRemaining, lockedUntil).
...
  responses:
    Unauthorized:
      description: Authentication missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Token inválido ou expirado.
            code: INVALID_TOKEN
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Conta não encontrada.
            code: ACCOUNT_NOT_FOUND
...
    ServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Erro interno do servidor.
            code: UNHANDLED_ERROR
paths:
...
  /expapi/v1/user/cancel-close-account:
    post:
      tags: [User]
      summary: Cancel scheduled account closure
      description: |
        Cancels a previously scheduled closure. Returns `400 NO_CLOSURE_SCHEDULED`
        if no closure is pending.

        CSRF-protected.
      operationId: cancelCloseAccount
      security:
        - bearerAuth: []
        - csrfToken: []
      parameters:
        - in: header
          name: X-CSRF-Token
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties: {}
      responses:
        '200':
          description: Closure cancelled.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  message:
                    type: string
        '400':
          description: No closure scheduled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Não há fechamento agendado para cancelar.
                code: NO_CLOSURE_SCHEDULED
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

  # ───────────────────────────────────────────────────────────────────────
  # PUBLIC PROFILE + NEWS
  # ───────────────────────────────────────────────────────────────────────
...
  /expapi/v1/admin/users/{userId}:
    put:
      tags: [Admin]
      summary: Update a user (support+)
      description: |
        Updates fields on a target user. The set of writable fields depends on
        the caller's level **and** the target's level (callers cannot modify
        peers or superiors):

        - Level ≥ 5 (target < 5): `blocked`
        - Level ≥ 6 (target < adminLevel): `banned`, `emailNotifications`,
          `discordNotifications`, `botActivityAlerts`
        - Level ≥ 7 (target < adminLevel, targetLevel < adminLevel,
          requested `accessType` < adminLevel): `accessType`
        - Level ≥ 8 (target < adminLevel): `firstName`, `lastName`,
          `emailVerified`, `publicProfile`, `showOnlineStatus`, `language`,
          `timezone`

        CSRF-protected.
      operationId: updateUser
      security:
        - bearerAuth: []
        - csrfToken: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: Target user's `accountId` (UUID).
        - in: header
          name: X-CSRF-Token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                blocked:
                  type: boolean
                banned:
                  type: boolean
                emailNotifications:
                  type: boolean
                discordNotifications:
                  type: boolean
                botActivityAlerts:
                  type: boolean
                accessType:
                  type: string
                  enum: [user, vipUser, enterpriseUser, contentCreator, tester, support,
                         moderator, admin, headadmin, developer, coowner, owner]
                firstName:
                  type: string
                lastName:
                  type: string
                emailVerified:
                  type: boolean
                publicProfile:
                  type: boolean
                showOnlineStatus:
                  type: boolean
                language:
                  type: string
                timezone:
                  type: string
            example:
              blocked: true
              banned: false
      responses:
        '200':
          description: User updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  updatedFields:
                    type: array
                    items:
                      type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Insufficient permission or no allowed fields for this caller/target.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                permission:
                  value:
                    error: Permissão insuficiente.
                    code: INSUFFICIENT_PERMISSION
                no_fields:
                  value:
                    error: Nenhum campo permitido para alteração.
                    code: NO_ALLOWED_FIELDS
        '404':
          description: Target user not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Usuário não encontrado.
                code: USER_NOT_FOUND
        '500':
          $ref: '#/components/responses/ServerError'
...